chore: move OIDC auth into own pkg
